requirements.txt
Flask
Flask-Cors
pydub
SpeechRecognition
requests
python-dotenv
gunicorn